move sidebar to its own feature
