+++ src/nvim/init.lua
@@ -5,6 +5,7 @@ local feature_languages = require('features.languages')
 local feature_scm = require('features.scm')
 local feature_lsp = require('features.lsp')
 local feature_completion = require('features.completion')
+local feature_sidebar = require('features.sidebar')
 
 local all_spec = {}
 vim.list_extend(all_spec, feature_editor.plugins)
@@ -12,6 +13,7 @@ vim.list_extend(all_spec, feature_languages.plugins)
 vim.list_extend(all_spec, feature_scm.plugins)
 vim.list_extend(all_spec, feature_lsp.plugins)
 vim.list_extend(all_spec, feature_completion.plugins)
+vim.list_extend(all_spec, feature_sidebar.plugins)
 
 -- Setup lazy.nvim
 require('lazy').setup({
